+++ src/CriticalTransitions.jl
@@ -62,7 +62,8 @@ include("noiseprocesses/stochprocess.jl")
 include("largedeviations/action.jl")
 include("largedeviations/min_action_method.jl")
 include("largedeviations/geometric_min_action_method.jl")
-include("RateSystem.jl")
+#include("RateSystem.jl")
+include("RateSystemDraft.jl")
 
 include("../systems/CTLibrary.jl")
 using .CTLibrary
@@ -80,7 +81,8 @@ export fw_action, om_action, action, geometric_action
 export min_action_method, geometric_min_action_method
 export make_jld2, make_h5, intervals_to_box
 export covariance_matrix, diffusion_matrix
-export RateSystem
+#export RateSystem
+export RateProtocol,RateSystem
 # export edgetracking, bisect_to_edge, AttractorsViaProximity
 # export fixedpoints
 # ^ extention tests needed
